Refactor structure of generator (#960)
