+++ pkg/generator/helper/helper_test.go
@@ -1,7 +1,6 @@
-package main
+package helper
 
 import (
-	//"bytes"
 	"fmt"
 	"io"
 	"io/ioutil"
