+++ pkg/generator/helper/testdata/TestProcessMetaFiles/step_code_generated.golden
