+++ pkg/generator/helper/testdata/TestProcessMetaFiles/test_code_generated.golden
